Move useragent to internal/useragent
